Shift logistic function in utils.photometry module and delete completeness_function module.
