+++ skypy/utils/__init__.py
@@ -7,7 +7,6 @@ __all__ = []
 from . import photometry  # noqa: F401
 from . import random  # noqa: F401
 from . import special  # noqa: F401
-from . import completeness_function  # noqa: F401
 
 from ._decorators import broadcast_arguments, dependent_argument
