+++ skypy/utils/tests/test_import.py
@@ -2,4 +2,3 @@ def test_import():
     import skypy.utils  # noqa: F401
     import skypy.utils.random  # noqa: F401
     import skypy.utils.special  # noqa: F401
-    import skypy.utils.completeness_function  # noqa: F401
